+++ wibble/singleton.h
@@ -74,7 +74,7 @@ public:
 	explicit Singleton(const T& value) : value(value) {}
 
 	bool empty() const { return false; }
-	unsigned int size() const { return 1; }
+	size_t size() const { return 1; }
 
 	iterator begin() { return iterator(&value); }
 	iterator end() { return iterator(); }
